Add files via upload
This assignment is a programming assignment wherein you have to build a multiple linear regression model for the prediction of demand for shared bikes.
Business Goal:
You are required to model the demand for shared bikes with the available independent variables. It will be used by the management to understand how exactly the demands vary with different features. They can accordingly manipulate the business strategy to meet the demand levels and meet the customer's expectations. Further, the model will be a good way for management to understand the demand dynamics of a new market.